Added "Show" command to Firmware, better explanations.
The s command now triggers an account to be displayed on the terminal,
instead of being typed. This also works with the search function by
pressing ENTER and then s.

The trigger function now tells what it does before doing it, so
if you select u%0f it will ask [U] Accountname ?
And if you type 0f it will ask [U][S][P] Accountname ?
And so on (s%0f) will ask [SHOW] Accountname ?