Added extra-paranoid-input option.
This option is only useful for entering the master-password,
but can double as "erase everything" feature.

When asked to enter a string, the "scamble" feature is turned off,
if the button is pressed, the counter is incremented,
the counter decides how the next characters will be scrambled.

If backspace is pressed when scramble is enabled, the whole string
is erased, and scramble is turned off.

Imagine you want a master password "mysafepasswordhere".
If you first enter "my" then press the button once
and enter "safe" then press the button once more
and enter "pass" and press the button 4 times "word"
and then three times before entering "here"
This will scramble the string before being passed on,
so if an attacker obtains the password, he will still need
to find out where, and how many times the button was pressed.
For this example, when unlocking the key requires you to type:
my, press once
safe, press once,
pass, press four times,
word, press three times
here.

This change is compatible with the existing storage layout, and can be
used on existing keys with no problem. It will not work with the GUI.