Add EEPROM check.
When a record is deleted, the random data is read back,
if the data is not identical to what was written, a Record with [Bad] record
is saved there to prevent user-data being written in the bad ara.

Format function use the delete function. This makes format
double as a crude EEPROM health check.