Better key derivation, random MAC.
The key derivation is improved over the previous implementation.

Instead of using the encrypted key as MAC, which potentially poses a weakness
in the AES encryption, a random MAC is generated.

This version lacks the macro function, and uses different text-strings to make
room for the SHA256 algorithm. It is not supported by the GUI or
backup tool. It is not compatible with existing FinalKey data.

It is for those who believe that they need more security than the standard
FinalKey implementation provides.